implemented trigger travel time supression
Full implementation of trigger travel time suppression.
The feature allows the user to set a delay in ms for before the controller passes active analog trigger inputs to the controlelr after they after they cross the lightshield threshold.
This is meant to prevent unwanted ATD when trying to powershield with a analog + digital trigger.

The setting shortcut is AY+L/R+DPAD up down and currently allow a delay of up to 16ms